add game revamp
