+++ src/components/components/NavTabs.js
@@ -18,6 +18,7 @@ import {
 	Toolbar,
 	useMediaQuery,
 } from '@mui/material';
+
 import { useTheme } from '@mui/styles';
 import {
 	CheckBox,
