+++ src/components/components/index.js
@@ -2,7 +2,6 @@ import AddGameModal from './AddGameModal';
 import AppBar from './AppBar';
 import NavTabs from './NavTabs';
 import ConfirmModal from './ConfirmModal';
-import GameList from './GameList';
 import GameCards from './GameCards';
 import LoadingOverlay from './Loading';
 import PlatformIcon from './PlatformIcon';
@@ -12,7 +11,6 @@ export {
 	AppBar,
 	NavTabs,
 	ConfirmModal,
-	GameList,
 	GameCards,
 	LoadingOverlay,
 	PlatformIcon,
